Improve the reliability of the rating process; explicit UTF-8
Specifying UTF-8 explicitly helps with edge cases when the CSV file is
created and read on different platforms (Linux vs Windows for example).

Some elements on the webpage can be slow to load because of the Internet
condition, adding extra EC conditions so that the actions can be
executed more reliably.